pull: Unify ccbynd with CC-BY-ND-4.0
[1] is associating 'CC-BY-ND' with "Creative Commons Noderivatives,
any version" and the 'non-free' tag.  This is a bit odd, because the
CC title for the BY-ND license is "Attribution-NoDerivatives" [2].
I'm not sure why the FSF left out "Attribution" here, but since the CC
doesn't have a ND-only license (all of their licenses besides CC0
include BY [3]), I'm assuming it's just FSF oversight.

[4] is associating 'ccbynd' with "Creative Commons
Attribution-NoDerivs 4.0 license (a.k.a. CC BY-ND)" and the
'viewpoint' tag.

[1]: https://www.gnu.org/licenses/license-list.en.html#CC-BY-ND
[2]: https://creativecommons.org/licenses/by-nd/4.0/
[3]: https://creativecommons.org/licenses/
[4]: https://www.gnu.org/licenses/license-list.en.html#ccbynd